refactor: centralize state management
