+++ src/spark-md5.d.ts
@@ -1,0 +1,4 @@
+declare module 'https://cdn.jsdelivr.net/npm/spark-md5@3.0.2/+esm' {
+  const SparkMD5: { hash: (str: string) => string }
+  export default SparkMD5
+}
