+++ src/storage/widgetStatePersister.js
@@ -1,0 +1,88 @@
+// @ts-check
+/**
+ * Utility for persisting the current state of widgets into the in-memory board structure,
+ * and then triggering a save to localStorage via the StorageManager.
+ *
+ * @module storage/widgetStatePersister
+ */
+import { Logger } from '../utils/Logger.js'
+import StorageManager from './StorageManager.js'
+
+const logger = new Logger('widgetStatePersister.js')
+
+/**
+ * Converts a widget DOM element into a serializable state object.
+ * @param {HTMLElement} widget - The widget element.
+ * @returns {import('../types.js').Widget} A serializable widget state object.
+ */
+function serializeWidgetState (widget) {
+  let metadata = {}
+  if (widget.dataset.metadata) {
+    try { metadata = JSON.parse(widget.dataset.metadata) } catch (e) { metadata = {} }
+  }
+
+  let settings = {}
+  if (widget.dataset.settings) {
+    try { settings = JSON.parse(widget.dataset.settings) } catch (e) { settings = {} }
+  }
+
+  return {
+    dataid: widget.dataset.dataid,
+    order: widget.getAttribute('data-order'),
+    url: widget.querySelector('iframe').src,
+    columns: widget.dataset.columns || '1',
+    rows: widget.dataset.rows || '1',
+    type: widget.dataset.type || 'iframe',
+    metadata,
+    settings
+  }
+}
+
+/**
+ * Serializes the state of all visible widgets for a given view,
+ * updates the in-memory board object, and then persists the entire
+ * board structure to localStorage via the StorageManager.
+ *
+ * @param {string} boardId - The ID of the current board.
+ * @param {string} viewId - The ID of the current view.
+ * @returns {void}
+ */
+export function saveWidgetState (boardId, viewId) {
+  if (!boardId || !viewId) {
+    return logger.error('Board ID or View ID is missing. Cannot save widget state.')
+  }
+
+  try {
+    // ALWAYS operate on the in-memory "live" state.
+    const board = window.asd.boards.find(b => b.id === boardId)
+    if (!board) return logger.error(`Board not found for saving state: ${boardId}`)
+
+    const view = board.views.find(v => v.id === viewId)
+    if (!view) return logger.error(`View not found for saving state: ${viewId}`)
+
+    const widgetContainer = document.getElementById('widget-container')
+    const visibleWidgets = Array.from(widgetContainer.children)
+      .filter(el => (el instanceof HTMLElement) && el.style.display !== 'none')
+
+    const sortedVisibleWidgets = visibleWidgets.sort((a, b) => {
+      const orderA = parseInt(/** @type {HTMLElement} */(a).getAttribute('data-order') || '0', 10)
+      const orderB = parseInt(/** @type {HTMLElement} */(b).getAttribute('data-order') || '0', 10)
+      return orderA - orderB
+    })
+
+    sortedVisibleWidgets.forEach((widget, index) => {
+      (/** @type {HTMLElement} */(widget)).setAttribute('data-order', String(index))
+      ;(/** @type {HTMLElement} */(widget)).style.order = String(index)
+    })
+
+    // Mutate the in-memory widgetState for the current view.
+    view.widgetState = sortedVisibleWidgets.map(widget => serializeWidgetState(/** @type {HTMLElement} */(widget)))
+
+    // Persist the entire updated boards structure.
+    StorageManager.setBoards(window.asd.boards)
+
+    logger.info(`Saved widget state for view: ${viewId} in board: ${boardId}`)
+  } catch (error) {
+    logger.error('Error saving widget state:', error)
+  }
+}
